requirements.txt
# --- Core AI & Data ---
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.1.0

# --- Vector Database ---
pymilvus>=2.3.0
milvus-lite>=2.3.0
rank-bm25>=0.2.2

# --- PDF Processing ---
pymupdf>=1.23.0
langdetect>=1.0.9

# --- LLM & API ---
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# --- Web App & UI ---
streamlit>=1.30.0
watchdog>=3.0.0

# --- Utilities ---
tqdm>=4.66.0
scikit-learn>=1.3.0